Added entities and data based on team chat

+++ Inventory_Management_6/src/main/java/sg/edu/iss/demo/repo/ReorderRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import sg.edu.iss.demo.model.Reorder;
 
-public interface ReorderRepository extends JpaRepository<Reorder, Long> {
+public interface ReorderRepository extends JpaRepository<Reorder, Integer> {
 
 }
